Use exact long source values in BCL proofs

+++ bcl/packages/multithreading/src/App.ts
@@ -1,6 +1,6 @@
 import { Console, Environment } from "@tsonic/dotnet/System.js";
 import { Parallel } from "@tsonic/dotnet/System.Threading.Tasks.js";
-import type { int, long } from "@tsonic/csharp/types.js";
+import type { long } from "@tsonic/csharp/types.js";
 
 export function main(): void {
   Console.WriteLine("=== Parallel Computation Test (BCL) ===");
@@ -8,7 +8,7 @@ export function main(): void {
   Console.WriteLine(`Processors: ${Environment.ProcessorCount}`);
   Console.WriteLine("");
 
-  const iterations: int = 100000;
+  const iterations: long = 100000n;
   Console.WriteLine(`Running 3 workers in PARALLEL with ${iterations} iterations each...`);
   Console.WriteLine("");
 
@@ -20,7 +20,7 @@ export function main(): void {
     () => {
       Console.WriteLine("Worker 1 starting on thread...");
       let sum: long = 0n;
-      for (let i: int = 0; i < iterations; i++) {
+      for (let i: long = 0n; i < iterations; i++) {
         sum += i;
       }
       Console.WriteLine(`Worker 1 done. Sum: ${sum}`);
@@ -29,7 +29,7 @@ export function main(): void {
     () => {
       Console.WriteLine("Worker 2 starting on thread...");
       let sum: long = 0n;
-      for (let i: int = 0; i < iterations; i++) {
+      for (let i: long = 0n; i < iterations; i++) {
         sum += i;
       }
       Console.WriteLine(`Worker 2 done. Sum: ${sum}`);
@@ -38,7 +38,7 @@ export function main(): void {
     () => {
       Console.WriteLine("Worker 3 starting on thread...");
       let sum: long = 0n;
-      for (let i: int = 0; i < iterations; i++) {
+      for (let i: long = 0n; i < iterations; i++) {
         sum += i;
       }
       Console.WriteLine(`Worker 3 done. Sum: ${sum}`);
